Move cutCellsHorizontallyAt() to general utils.
